Add AnnotateModels::FilePatterns
